Computer Vision/segmenting/sam3/requirements.txt
numpy
Pillow
python-dotenv
torch
torchvision
git+https://github.com/facebookresearch/sam3.git
